% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package load hook}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The path to the library where the package is installed.}

\item{pkgname}{The name of the package.}
}
\description{
This function is automatically called when the package is loaded.
It registers global variables to avoid R CMD check notes about
no visible binding for global variables.
}
\keyword{internal}
