% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_helpers.R
\name{get_mode}
\alias{get_mode}
\title{Find the Most Common Value (Mode)}
\usage{
get_mode(x)
}
\arguments{
\item{x}{A vector of values (numeric, character, factor, etc.)}
}
\value{
The most frequently occurring value in \code{x}.
}
\description{
Returns the statistical mode of a vector (first in case of ties).
}
