% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_WORLD.R
\name{hello_WORLD}
\alias{hello_WORLD}
\title{Simple Hello World function}
\usage{
hello_WORLD()
}
\description{
Returns "Hello, World!" in English
}
\details{
Hello World in English
}
\examples{
hello_WORLD()  # Returns English greeting

}
