% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_first_entered.R
\name{make_first_entered}
\alias{make_first_entered}
\title{Get First Entered}
\usage{
make_first_entered(dataset)
}
\arguments{
\item{dataset}{A data frame that includes student grade level and academic year}
}
\value{
A data frame with an appended column containing student entry cohort data
}
\description{
Appends a column of "FIRST_ENTERED" to the dataset to indicate which cohort they were initially included in
}
\details{
Append a Column for Grade Level and Year indicating when the Student first appeared in the Dataset
}
\examples{
make_first_entered(dataset = math)

}
