% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_cohort.R
\name{parallel_cohort}
\alias{parallel_cohort}
\title{Parallel Cohort Analysis}
\usage{
parallel_cohort(YEARS, START_GRADE, CONTENT)
}
\arguments{
\item{YEARS}{Vector indicating which years to include in the analysis}

\item{START_GRADE}{Integer indicating which grade level to start with}

\item{CONTENT}{Character string indicating which content area to work with}
}
\value{
A list containing the greeting and optional metadata
}
\description{
Makes a plot of parallel cohorts
}
\details{
Perform Parallel Cohort Analysis
}
