% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alluvial_mobility.R
\name{plot_alluvial_mobility}
\alias{plot_alluvial_mobility}
\title{Plot Alluvial Mobility}
\usage{
plot_alluvial_mobility(
  dataset,
  start_year,
  start_grade,
  print_table = TRUE,
  make_plot = TRUE
)
}
\arguments{
\item{dataset}{A data frame that includes student grade level and academic year}

\item{start_year}{The academic year to begin evaluating with}

\item{start_grade}{The grade level to begin evaluating with}

\item{print_table}{Prints a contingency table of indicators associated with student mobility (default: TRUE)}

\item{make_plot}{Makes an alluvial diagram depicting student mobility (default: TRUE)}
}
\value{
A table of proportions of students
}
\description{
Creates an Alluvial Diagram depicting Student Mobility from previous year
}
\details{
Plot an Alluvial Diagram of Student Mobility
}
