% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cohorts.R
\name{get_cohorts}
\alias{get_cohorts}
\title{Split Long Dataset into Cohorts by Grade and Year}
\usage{
get_cohorts(dataset)
}
\arguments{
\item{dataset}{A data frame in long format with at least two columns: \code{GRADE} and \code{YEAR}.}
}
\value{
A list of data frames, each corresponding to a unique grade-year combination (cohort).
}
\description{
This function takes a long-format dataset with grade and year information
and splits it into a list of cohort-specific datasets.
}
\examples{
get_cohorts(dataset = math)

}
