% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_COHORTS.R
\name{get_COHORTS}
\alias{get_COHORTS}
\title{Get Cohort Data from Long Dataset}
\usage{
get_COHORTS(DATASET)
}
\arguments{
\item{DATASET}{a data frame in long formate with columns for GRADE and YEAR}
}
\value{
A list containing cohort datasets by grade level and year
}
\description{
Returns "Hello, World!" in various languages using the Hello_World dataset
}
\details{
Takes a long dataset and separates it into a list of cohort datasets by grade level and year
}
