% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mobility.R
\name{make_mobility}
\alias{make_mobility}
\title{Make Mobility}
\usage{
make_mobility(
  dataset,
  start_year,
  start_grade,
  print_table = TRUE,
  make_plot = TRUE
)
}
\arguments{
\item{dataset}{A data frame that includes student grade level and academic year}

\item{start_year}{The academic year to begin evaluating with}

\item{start_grade}{The grade level to begin evaluating with}

\item{print_table}{Prints a table of proportions of students for student mobility (default: TRUE)}

\item{make_plot}{Makes a relative frequency bar plot of student mobility (default: TRUE)}
}
\value{
A table of proportions of students and an alluvial
}
\description{
Appends a year-to-year student mobility category based on a given starting year
}
\details{
Append a Factor for Student Mobility
}
\examples{
make_mobility(dataset = math, start_year = "2019_2020", start_grade = 3)
make_mobility(dataset = math, start_year = 2021, start_grade = 6, print_table = FALSE)

}
