% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proficiency_levels.R
\name{make_proficiency_levels}
\alias{make_proficiency_levels}
\title{Get Proficiency}
\usage{
make_proficiency_levels(
  dataset,
  n_proficiencies,
  print_table = TRUE,
  make_plot = TRUE
)
}
\arguments{
\item{dataset}{A data frame that includes student achievement levels}

\item{n_proficiencies}{An integer for the number of achievement levels that are considered proficient}

\item{print_table}{Prints a table of proportions of students who are proficient and not proficient (default: TRUE)}

\item{make_plot}{Makes a relative frequency bar plot of students who are proficient and not proficient (default: TRUE)}
}
\value{
A data frame with an appended column containing student proficiency data
}
\description{
Appends a student proficiency data based on the achievement level and the number of achievement levels that are considered proficient
}
\details{
Append a Binary Factor for Student Proficiency
}
\examples{
make_proficiency_levels(dataset = math, n_proficiencies = 2)

}
