% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alluvial_mobility.R
\name{plot_alluvial_mobility}
\alias{plot_alluvial_mobility}
\title{Plot Alluvial Mobility}
\usage{
plot_alluvial_mobility(dataset, start_year, start_grade, print_table = TRUE)
}
\arguments{
\item{dataset}{A data frame that includes student grade level and academic year}

\item{start_year}{The academic year to begin evaluating with (only required if not already used make_mobility)}

\item{start_grade}{The grade level to begin evaluating with (only required if not already used make_mobility)}

\item{print_table}{Prints a contingency table of indicators associated with student mobility (default: TRUE)}
}
\value{
An Alluvial Diagram showing Gender, Ethnicity, and Mobility with Optional Contingency Table
}
\description{
Creates an Alluvial Diagram depicting Student Mobility from previous year
}
\details{
Plot an Alluvial Diagram of Student Mobility
}
\examples{
plot_alluvial_mobility(dataset = math, start_year = "2019_2020", start_grade = 3, print_table = FALSE)

}
